test(agent): classify provider bridge contracts as integration tests
